the Search All Missing button was searching all missing and monitored only though the dialog that popped up was informing the user it was going to search for x movies, where x corresponded to the number of movies filtered on the page.
I changed this button, so now it will search all the items as they are
filtered on the page.

For example, if you want to search all missing (regardless of
monitor/unmonitor) click the all filter and click the button.

If you want to search only monitored/missing, click the monitored button
and then click the search all button...

this included the old functionality, but allows the user alot more
flexibility...

i also added the all filter, and refactored the code,
so that builds the expression for the LINQ.. since this needed to be
used in two places.. just implement it once and use it in both places..

I tested this out... and stepped through with debugger.. i also did a
quick test of everything else. Im confident that the featureset
implemented and bugs fixed by this commit are OK... Im not 100% that
other parts of radarr dont use the same MissingMovieSearch routines..
but i dont think so...