macOS chmod fix for Radarr executable file
This is a possible fix for issue #745. It builds upon AeonLucid’s commit hash aace29e1e14dd142a4a6806682c55f4bcdba81ba (#1539), which was unfortunately not the solution to making Radarr.app/Contents/MacOS/Radarr be executable. Although I’m not familiar with appveyor, after looking at the code, my best guess is that moving the cygwin-powered chmod command from the PackageOsx Task to the PackageOsxApp Task may be the solution. Not sure, but thought I’d give it a shot. Also, this is my first ever pull request on a public github project, so, hope I did everything correctly.