Add configurable setting for "replace illegal characters" behavior
The default functionality is to delete colon characters. However,
there are several other ways that various media software expects
these characters to be resolved (dash, space dash, or space dash
space). This commit adds the ability to choose one of those
settings from a "Colon Action" select control in the "Movie Naming"
section of the "Media Management" Settings tab.

Also, this changes behavior for the "Rename Movies" switch so that
when it is disabled it also hides the "Replace Illegal Characters"
setting. This is in line with how these two settings interoperate.

I only resolved two issues with a single commit because these both
modify the same UI elements.