Don't return unmapped folders on rootfolder API call
The unmapped folders returned by the "rootfolder" API call are simply
not used at all. Folder bulk import returns them in an extra call.

Beside this, retrieving them with the current implementation takes
~410ms per rootfolder on my end. This is mainly because of the call
for `_movieRepository.All().ToList()`, which stored instead of calling
it for every rootfolder. But even then it still takes ~60ms.

This might not be much for users with a single rootfolder, but for me
this sums up to 8 seconds for this simple API request (or up to a full
minute without the list-fix).