Added Dolby Digital Plus with Atmos
Added Dolby Digital Plus with Atmos by checking the codec additional features for JOC, which stands for Joint Object Coding and is the technique used for adding Atmos metadata to Dolby Digital Plus audio streams. This is similar to the 16-ch additional feature used to distinguish Atmos in a TrueHD audio stream.