Fix bug: Parsing HTML encoded ampersands in magnet urls
For HTML encoded magnet urls, if the ampersand is encoded, the HTML
decoding will still reveal an ampersand, so we need to replace those.
Use of a regex because we need to only change parts of the display
name, as query parameters notation also use ampersands.
Only decode HTML for display name part of URL and replace resulting
ampersand characters with specific string "And".