Add Custom Format condition for Indexer
This commit implements a new Custom Format condition to match on the specific Indexer that was used to grab the content. The Indexer ID is the property used for the comparision (and not the Indexer Name), however, the condition editor GUI displays the Indexer Name for a more user friendly experience (instead of the internal Indexer ID).
The ability to store in the database the Indexer ID was also added to Blocklist and MovieFile items. The Indexer ID is also exposed and tracked in all applicable History events.