Update SkyHookProxy.cs to add OriginalTitle to AlternativeTitles
The OriginalTitle property is not being used by Radarr, which limits its search results. It only uses "Title" and "AlternativeTitles", when usually in the case of non-english movies, the OriginalTitle is the one likely to get the most search results. So with this modification, OriginalTitle will also be appended the the AlternativeTitles list.